New structure for lib api
