+++ scraper_factory/__init__.py
@@ -1,2 +1,2 @@
-from scraper_factory.core.scrape import scrape
+from scraper_factory.core.api import scrape, list
 from scraper_factory.core.spidermanager import SpiderManager
